// lib/generated/prisma/models/District.ts
export type DistrictCreateInput = {
  id?: string
  title: string
  description?: string | null
  createdAt?: Date | string
  updatedAt?: Date | string
  House?: Prisma.HouseCreateNestedManyWithoutDistrictInput
}